pyproject.toml
[tool.poetry]
name = "mlvision"
version = "0.1.0"
description = ""
authors = ["Guichard Victor <guichardvictor@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.0.1"
torchvision = "^0.15.2"
pyyaml = "^6.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
ruff = "^0.0.278"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
